added database.py for Flask Blueprint
