+++ database.py
@@ -1,0 +1,6 @@
+from flask_login import LoginManager
+from flask_sqlalchemy import SQLAlchemy
+
+db = SQLAlchemy()
+login_db = SQLAlchemy()
+login_manager = LoginManager()
