+++ dcc-dashboard-service.wsgi
@@ -1,5 +1,8 @@
 import sys
 import os
+import logging
+
+logging.basicConfig()
 
 activate_this = '/var/www/html/dcc-dashboard-service/env/bin/activate_this.py'
 execfile(activate_this, dict(__file__=activate_this))
@@ -14,6 +17,6 @@ def application(req_environ, start_response):
     for key in ENV_VAR:
        os.environ[key] = req_environ.get(key, '')
     ## has to import my app inside of application def block.
-    from mapi import app as _application
+    from mapi import mapi as _application
 
     return _application(req_environ, start_response)
