+++ uwsgi/uwsgi.ini
@@ -1,4 +1,4 @@
 [uwsgi]
-module = mapi
+module = mapi:create_app()
 callable = app
 logto = /app/log/error_uwsgiB.log
